\name{Booth2}
\docType{data}
\alias{Booth2}
\title{A Logit-Normal GLMM Dataset}
\description{This data set contains simulated data from the paper of Booth and Hobert (referenced below) as well as another vector. }
}
\usage{data(BoothHobert)}
\format{
  A data frame with 3 columns:

   \describe{
    \item{y}{Response vector.}
    \item{x1}{Fixed effect model matrix. The matrix has just one column vector.}
    \item{z1}{A categorical vector to be used for part of the random effect model matrix.}
    \item{z2}{A categorical vector to be used for part of the random effect model matrix.}
  }
}
\details{The original data set was generated by Booth and Hobert using a single variance component, a single fixed effect, no intercept, and a logit link. This data set has the z2 vector added purely for illustration purposes.}

\references{Booth, J. G. and Hobert, J. P. (1999). Maximizing generalized linear mixed model likelihoods with an automated Monte Carlo EM algorithm. Journal of the Royal Statistical Society Series B (Statistical Methodology) 61, 265–285. }
\examples{
data(Booth2)
}
\keyword{generalized linear mixed models}

