% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervals.R
\name{ci}
\alias{ci}
\title{Compute Confidence Interval}
\usage{
ci(data, example = "epilepsy", B, seed = NULL)
}
\arguments{
\item{data}{The epilepsy dataframe with id, age, expind, treat and seizure columns}

\item{example}{The name of the dataframe - in this scenario it is "epilepsy"}

\item{B}{Numeric value to represent the number of times boostrapping should be done}

\item{seed}{optional argument to set a seed before bootstrapping so that results can be replicated}
}
\value{
A list of confidence intervals for each variable in the dataset
}
\description{
Compute Confidence Interval
}
\examples{
ci(data = epilepsy,example = "epilepsy",B = 20,seed = 1)
ci(data = epilepsy,example = "epilepsy",B = 20)
}
