% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMM-package.R
\docType{package}
\name{GLMM-package}
\alias{GLMM}
\alias{GLMM-package}
\title{GLMM: Fitting a Generalized Linear Mixed Model for a Poisson Distribution}
\description{
This package fits a Poisson GLMM and provides estimates of coefficients, standard errors and confidence intervals using bootstrapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tito1224/GLMM}
  \item Report bugs at \url{https://github.com/tito1224/GLMM/issues}
}

}
\author{
\strong{Maintainer}: Angela Zheng \email{azheng26@uwo.ca}

Authors:
\itemize{
  \item Oluwatitomi Adebajo \email{oadebajo@uwo.ca}
}

}
\keyword{internal}
