% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglmm-methods.R
\name{ranef}
\alias{ranef}
\alias{ranef.fastglmm}
\title{Random-effect predictions (not available)}
\usage{
ranef(object, ...)

\method{ranef}{fastglmm}(object, ...)
}
\arguments{
\item{object}{a \link{fastglmm} fit.}

\item{...}{unused.}
}
\description{
Errors: conditional modes are computed inside the kernel but not surfaced
on \code{glmm::Fit} yet. The engine spec
\code{2026-07-15-engine-loglik-diagnostics} is what lifts this.
}
