% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{run_model}
\alias{run_model}
\title{Final Project Model Fitting}
\usage{
run_model(data, example = "tortoise")
}
\arguments{
\item{data}{the data set for your example}

\item{example}{the name of your example: one of "culcita", "ctsib", "epilepsy", or "tortoise"}
}
\value{
A list with summary statistics from the fitted model.
}
\description{
Final Project Model Fitting
}
\examples{
##### Epilepsy #####

## Fit model to epilepsy data. Fixed effects in the model are:
##  (Intercept) -- the intercept
##  age -- age as a continuous predictor
##  expind -- a categorical variable with two levels (0 for before and 1 for after)
##  expind:treat -- a categorical variable with two levels (1 for
##  observations from individuals on the drug in the after period and 0 otherwise)

epilepsy_fit <- run_model(epilepsy, "epilepsy")


}
