% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btsp.R
\name{btsp}
\alias{btsp}
\title{Conduct bootstrapping}
\usage{
btsp(data, example = "epilepsy", B, seed = NULL)
}
\arguments{
\item{data}{The epilepsy dataframe with id, age, expind, treat and seizure columns}

\item{example}{The name of the dataframe - in this scenario it is "epilepsy"}

\item{B}{Numeric value to represent the number of times boostrapping should be done}

\item{seed}{optional argument to set a seed before bootstrapping so that results can be replicated}
}
\value{
A list of standard errors for each variable in the dataset
}
\description{
Conduct bootstrapping
}
\examples{
btsp(data = epilepsy,example = "epilepsy",B = 20,seed = 1)
btsp(data = epilepsy,example = "epilepsy",B = 20)
}
