% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervals.R
\name{ci}
\alias{ci}
\title{Compute Confidence Interval}
\usage{
ci(data, example = "epilepsy", B, seed = NULL)
}
\arguments{
\item{data}{The epilepsy data set with id, age, expind, treat and seizure columns}

\item{example}{The name of the data set (called "epilepsy" in this case)}

\item{B}{Numeric value to represent the number of times bootstrapping should be done}

\item{seed}{Optional argument to set a seed before bootstrapping so that results can be saved}
}
\value{
A list of confidence intervals for each variable in the data set
}
\description{
Compute Confidence Interval
}
\examples{
ci(data = epilepsy,example = "epilepsy",B = 20,seed = 1)
ci(data = epilepsy,example = "epilepsy",B = 20)
}
