\name{glmm-package}
\alias{glmm-package}
\alias{glmm}
\docType{package}
\title{Generalized Linear Mixed Models using Monte Carlo Likelihood Approximation }
\description{Approximates the likelihood of a GLMM using Monte Carlo likelihood approximation.  Maximizes the approximated likelihood to return maximum likelihood estimates and other model information.}
\details{
\tabular{ll}{
Package: \tab glmm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-26\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Christina Knudson

Maintainer: Christina Knudson <christina@stat.umn.edu>}
\references{
Geyer, C. J. and Thompson, E. (1992). Constrained Monte Carlo maximum likelihood for dependent data. Journal of the Royal Statistical Society, Series B, 54:657–699.

Geyer, C. J. (1994). On the convergence of Monte Carlo maximum likelihood calculations. Journal of the Royal Statistical Society, Series B, 61:261–274.

Sung, Y. J. and Geyer, C. J. (2007). Monte Carlo likelihood inference for missing data models. Annals of Statistics, 35:990–1011.


}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
