% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglmm-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}}

