% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{run_model}
\alias{run_model}
\title{Final Project Model Fitting}
\usage{
run_model(data, example = "tortoise")
}
\arguments{
\item{data}{the data set for your example}

\item{example}{the name of your example: one of "culcita", "ctsib", "epilepsy", or "tortoise"}
}
\value{
A list with summary statistics from the fitted model.
}
\description{
Final Project Model Fitting
}
\examples{
##### Epilepsy #####

## Load data
epilepsy <- read_csv("epilepsy.csv")

## Sum separate observations for each patient in the after period
epilepsy <- epilepsy \%>\%
  group_by(id,treat,expind,age) \%>\%
  summarize(seizures = sum(seizures),
            .groups = "drop")

## Fit model to epilepsy data. Fixed effects in the model are:
##  (Intercept) -- the intercept
##  age -- age as a continuous predictor
##  expind -- a categorical variable with two levels (0 for before and 1 for after)
##  expind:treat -- a categorical variable with two levels (1 for observations from individuals on the drug in the after period and 0 otherwise)

epilepsy_fit <- run_model(epilepsy, "epilepsy")

##### CTSIB #####

## Load data
ctsib <- read_csv("ctsib.csv")

## Define response (stable or not)
ctsib <- ctsib \%>\%
  mutate(stable = 1 * (CTSIB == 1))

## Fit model to ctsib data. Fixed effects in the model include:
##  (Intercept) -- the intercept
##  Surface -- a categorical variable with two levels (foam and norm)
##  Vision -- a categorical variable with three levels (closed, dome, open)

ctsib_fit <- run_model(ctsib, "ctsib")

##### Tortoise #####
## Load data
tortoise <- read_csv("gopher_tortoise.csv")

## Fit model to tortoise data. Fixed effects in the model are:
##   (Intercept) -- the intercept
##   prev -- the seroprevalence
##   year -- year effect (as a categorical variable)

tortoise_fit <- run_model(tortoise, "tortoise")

##### Culcita #####
## Load data
culcita <- read_csv("culcita.csv")

## Fit model to culcita data. Fixed effects in the model are:
##   (Intercept) -- the intercept
##   ttt -- the treatment as a categorical variable with four levels: none, crabs, shrimp, or both

culcita_fit <- run_model(culcita, "culcita")
}
